Updated OSX project file, fixed some minor compile warnings, and bumped version # to 4.5 for the next release (whenever it may happen).
git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@3057 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba

+++ src/common/PNGLibrary.cxx
@@ -307,7 +307,7 @@ void PNGLibrary::loadImagetoSurface(FBSurface& surface)
 void PNGLibrary::writeComments(png_structp png_ptr, png_infop info_ptr,
                                const VariantList& comments)
 {
-  uInt32 numComments = comments.size();
+  uInt32 numComments = (int)comments.size();
   if(numComments == 0)
     return;
