+++ src/gui/ListWidget.cxx
@@ -144,7 +144,7 @@ const string& ListWidget::getSelectedString() const
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 void ListWidget::scrollTo(int item)
 {
-  int size = _list.size();
+  int size = (int)_list.size();
   if (item >= size)
     item = size - 1;
   if (item < 0)
@@ -160,7 +160,7 @@ void ListWidget::scrollTo(int item)
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 void ListWidget::recalc()
 {
-  int size = _list.size();
+  int size = (int)_list.size();
 
   if (_currentPos >= size)
     _currentPos = size - 1;
@@ -172,7 +172,7 @@ void ListWidget::recalc()
 
   _editMode = false;
 
-  _scrollBar->_numEntries     = _list.size();
+  _scrollBar->_numEntries     = (int)_list.size();
   _scrollBar->_entriesPerPage = _rows;
 
   // Reset to normal data entry
@@ -369,7 +369,7 @@ bool ListWidget::handleEvent(Event::Type e)
     case Event::UIPgDown:
       _selectedItem += _rows - 1;
       if (_selectedItem >= (int)_list.size() )
-        _selectedItem = _list.size() - 1;
+        _selectedItem = (int)_list.size() - 1;
       break;
 
     case Event::UIHome:
@@ -377,7 +377,7 @@ bool ListWidget::handleEvent(Event::Type e)
       break;
 
     case Event::UIEnd:
-      _selectedItem = _list.size() - 1;
+      _selectedItem = (int)_list.size() - 1;
       break;
 
     case Event::UIPrevDir:
@@ -444,7 +444,7 @@ void ListWidget::scrollToCurrent(int item)
   if (_currentPos < 0 || _rows > (int)_list.size())
     _currentPos = 0;
   else if (_currentPos + _rows > (int)_list.size())
-    _currentPos = _list.size() - _rows;
+    _currentPos = (int)_list.size() - _rows;
 
   int oldScrollPos = _scrollBar->_currentPos;
   _scrollBar->_currentPos = _currentPos;
